+++ Changelog.md
@@ -8,6 +8,7 @@
 - Added `Taxon::findOneByParentsAndSlug` method
 - Added `Taxonomy::findOneBySlug` method
 - Added PHP 8 support
+- Added `taxa` (and `taxons` alias) relationship to `Taxonomy` model
 
 ### 2.0.0
 ##### 2020-10-11